pyproject.toml
[project]
name = "stocktracker-api"
version = "0.0.1"
description = "Stock Tracker API - A FastAPI application for tracking stocks"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]",
    "uvicorn[standard]",
    "sqlmodel",
    "asyncpg",
    "alembic",
    "pydantic-settings",
    "passlib[bcrypt]",
    "python-jose[cryptography]",
    "python-keycloak",
    "python-multipart",
    "pydantic[email]>=2.12.5",
    "structlog>=24.4.0",
    "colorama>=0.4.6",
    "redis>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "pyright",
    "pytest",
    "pytest-asyncio",
    "httpx",
    "pre-commit",
    "factory-boy",
]

[tool.ruff]
target-version = "py312"
line-length = 120
src = ["app", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "SIM",    # flake8-simplify
    "T20",    # flake8-print
    "RUF",    # ruff-specific rules
]
ignore = [
    "B008",   # do not perform function calls in argument defaults (needed for Depends)
]

[tool.ruff.lint.isort]
known-first-party = ["app", "tests"]

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "basic"
venvPath = "."
venv = ".venv"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning",
]
markers = [
    "unit: pure-logic tests with no I/O",
    "integration: tests that exercise the HTTP layer with stubbed dependencies",
    "e2e: end-to-end tests against a live server",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
